requirements-pc-cuda.txt
# this file contains package requirements list for linux/windows PC with CPU and CUDA support

--extra-index-url https://download.pytorch.org/whl/cu118

numpy>=1.24.3
opencv-python>=4.8.0.74
onnx==1.14.0
insightface==0.7.3
psutil==5.9.5
tk==0.1.0
customtkinter==5.2.0
pillow==10.0.0
protobuf>=4.24.1
tqdm>=4.65.0
gfpgan==1.3.8
tensorflow>=2.13.0
future==0.18.3
colorama>=0.4.6
sympy==1.12
torch==2.0.1+cu118
torchvision==0.15.2+cu118
onnxruntime==1.15.1
onnxruntime-gpu==1.15.1
pyvirtualcam
pathvalidate
pygame